+++ src/scripts/landing/components/atoms/Confirmation.tsx
