+++ src/scripts/landing/components/organism/Lead.tsx
