+++ src/scripts/landing/index.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import * as ReactDOM from 'react-dom'
-import App from './App'
+import App from './components/templates/App'
 import {Language} from '../../i18n/landing'
 
 const lang = (() => {
